Improves user data handling and validation
Enhances user data access security by hiding password by default.

Adds validation for user IDs and improves error handling for invalid IDs.

Refactors educator creation to streamline data processing and improve error management.

+++ src/data-access/user.data-access.js
@@ -1,4 +1,9 @@
+import createError from "http-errors";
+import mongoose from "mongoose";
+
 import { UserModel } from "#models/index.js";
+
+const { isValidObjectId } = mongoose;
 
 export const user = {
   save: {
@@ -14,7 +19,7 @@ export const user = {
 
   read: {
     allUsers: async () => {
-      return await UserModel.find().select("-password");
+      return await UserModel.find();
     },
 
     userByEmail: async (email) => {
@@ -24,12 +29,20 @@ export const user = {
     },
 
     userById: async (id) => {
-      return await UserModel.findById(id).select("-password");
+      if (!isValidObjectId(id)) {
+        throw createError(400, "Invalid user ID format.");
+      }
+
+      return await UserModel.findById(id);
     },
   },
 
   update: {
     userById: async (id, userData) => {
+      if (!isValidObjectId(id)) {
+        throw createError(400, "Invalid user ID format.");
+      }
+
       return await UserModel.findByIdAndUpdate(id, userData, {
         new: true,
         upsert: true,
@@ -47,13 +60,17 @@ export const user = {
       return await UserModel.findOneAndUpdate(
         { email },
         { password },
-        { new: true, upsert: true },
+        { new: true, upsert: true }
       );
     },
   },
 
   remove: {
     userById: async (id) => {
+      if (!isValidObjectId(id)) {
+        throw createError(400, "Invalid user ID format.");
+      }
+
       return await UserModel.findByIdAndDelete(id);
     },
   },
